% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_gld_validity.R
\name{checkGLDValidN}
\alias{checkGLDValidN}
\title{Check if many GLDs are valid}
\usage{
checkGLDValidN(lambdas)
}
\arguments{
\item{lambdas}{lambda values of multiple GLD functions.}
}
\description{
This function check if many GLDs are valid.
}
\examples{
checkGLDValidN(lambdas)
}
\keyword{lambdas}
