% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_fit.R
\name{createDataset}
\alias{createDataset}
\title{Create a multidimentional array with simulations form a set of .csv.}
\usage{
createDataset(directory, pattern, dimension)
}
\arguments{
\item{directory}{the directory where the .csv files are stored.}

\item{pattern}{a pattern to read the .csv (e.g. pattern="^[h]").}

\item{dimension}{the dimensions to be used, NxMxS.}
}
\description{
This function read all the csv that are stored in a directory and create an NxMxS array, where N and M are sparial dimensions
and S is the number of simulations on each spatial point.
}
\examples{
directory = "~/PhD/thesis_phd/python_codes/datasets/20"
pattern="^[h]"
dimension = c(60, 60, 500)
mArray <- createDataset(directory, pattern, dimension)

}
\author{
Noel Moreno Lemus
}
