% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{createSyntheticData}
\alias{createSyntheticData}
\title{Create a synthetic dataset to test the clustering algorithm using the GLD}
\usage{
createSyntheticData(k = 11, d = 1, n = 1450, s = 1000)
}
\arguments{
\item{k}{number of clusters to test.}

\item{d}{dimension (e.g. if we have temperatute and preasure d = 2)}

\item{n}{the number of objects of the data}

\item{s}{the sample size}
}
\description{
This function create a synthetic dataset using the approach proposed in B. Jiang, J. Pei, Y. Tao and X. Lin, "Clustering Uncertain Data Based on Probability Distribution Similarity," in IEEE Transactions on Knowledge and Data Engineering, vol. 25, no. 4, pp. 751-763, April 2013. doi: 10.1109/TKDE.2011.221
}
\examples{
dataset <- createSyntheticData(3, 2, 1000, 500)

}
\references{
B. Jiang, J. Pei, Y. Tao and X. Lin, "Clustering Uncertain Data Based on Probability Distribution Similarity," in IEEE Transactions on Knowledge and Data Engineering, vol. 25, no. 4, pp. 751-763, April 2013.
doi: 10.1109/TKDE.2011.221
}
\author{
Noel Moreno Lemus
}
