% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utilities.R
\name{distGLDComparisonKL}
\alias{distGLDComparisonKL}
\title{Function to be used as a distance function in clustering algoritms}
\usage{
distGLDComparisonKL(x, centers)
}
\arguments{
\item{x}{dataset.}

\item{centers}{the centroids to be analized.}
}
\description{
This function return the distances between all the centroids and the elements of the dataset, using KL-divergence as a measure of the distance.
}
\examples{
TODO

}
\author{
Noel Moreno Lemus
}
