% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utilities.R
\name{gldClusterComparison}
\alias{gldClusterComparison}
\title{Evaluate if the centroid of a cluster is a good representative of the other members of this cluster}
\usage{
gldClusterComparison(cluster_number, n, centroid)
}
\arguments{
\item{cluster_number}{The cluster ID we are interesting in.}

\item{n}{Number of elements of the cluster to analize.}

\item{centroid}{The lambda values of the centroid of the cluster.}
}
\description{
This function return a list of D distances from the ks-test that compare the centroid of a cluster wiht n members of this cluster.
}
\examples{

Ds <- gldClusterComparison(3, 60, c(0, 2, 1.5, 1.3))

}
\author{
Noel Moreno Lemus
}
