% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utilities.R
\name{gldComparison}
\alias{gldComparison}
\title{Compare if two GLDs belongs to the same distribution}
\usage{
gldComparison(L1, L2)
}
\arguments{
\item{L1}{Lambda values of the first GLD.}

\item{L2}{Lambda values of the second GLD.}
}
\description{
This function return D distance from the ks-test that compare if both GLDs belongs to the same distribution.
}
\examples{
L1 = c(0, 2, 0.25, 1.5)
L2 = c(0, 2, 0.3, 1.75)
D <- gldComparison(L1, L2)

}
