% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utilities.R
\name{gldComparisonKL}
\alias{gldComparisonKL}
\title{Compare if two GLDs belongs to the same distribution based in its KL.dist}
\usage{
gldComparisonKL(L1, L2, param = "fmkl", no.test = 1000, len = floor(0.9 *
  no.test), alpha = 0.05)
}
\arguments{
\item{L1}{Lambda values of the first GLD.}

\item{L2}{Lambda values of the second GLD.}
}
\description{
This function return the KL.dist of two GLDs.
}
\examples{
L1 = c(0, 2, 0.25, 1.5)
L2 = c(0, 2, 0.3, 1.75)
D <- gldComparisonKL(L1, L2)

}
\author{
Noel Moreno Lemus
}
