% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_fit.R
\name{gldfitLmoments}
\alias{gldfitLmoments}
\title{Compute the lambda values of the GLD using the method of the Lmoments}
\usage{
gldfitLmoments(data)
}
\arguments{
\item{data}{The raw data to fit.}
}
\description{
This function allows you to show image with scale.
}
\examples{
gldfitLmoments(data)
}
\keyword{lmoments}
