% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_fit.R
\name{gldfitMoments}
\alias{gldfitMoments}
\title{Compute the lambda values of the GLD using the method of Moments}
\usage{
gldfitMoments(data)
}
\arguments{
\item{data}{The raw data to fit.}
}
\description{
This function allows you to show image with scale.
}
\examples{
gldfitMoments(data)
}
\keyword{lmoments}
