% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_fit.R
\name{gldFitParallel}
\alias{gldFitParallel}
\title{Fit the GLD to a multidimensional array in parallel.}
\usage{
gldFitParallel(nodes, dimension, mArray)
}
\arguments{
\item{nodes}{number of nodes to be used in the computation.}

\item{dimension}{the dimensions to be used, NxMxS.}

\item{mArray}{a multidimentional array with dimensions @param dimension}
}
\description{
This function compute the GLD that best fit each distribution on each spatial location.
}
\examples{
mArray <- createDataset(directory, pattern, dimension)
result <- gldFitParallel(4, dimension = c(60, 60), mArray = mArray)

}
\author{
Noel Moreno Lemus
}
