% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gld_utilities.R
\name{gldPlot}
\alias{gldPlot}
\title{Plot the GLD based in the Lambda values}
\usage{
gldPlot(L)
}
\arguments{
\item{L}{Lambda values of the GLD.}
}
\description{
This function allows you to show image with scale.
}
\examples{
L = c(0, 2, 0.25, 1.5)
gldPlot(L)

}
\keyword{lmoments}
